+++ src/wnm/__init__.py
@@ -1,3 +1,3 @@
 """A service to manage a cluster of decentralized Autonomi nodes"""
 
-__version__ = "0.0.29"
+__version__ = "0.0.30"